Deduplicated spinner logic into tryImportTemplate
